Add form field to html
Added a form field and button to html to allow for variable number of squares.
Also added an event listener and modifed and added some functions to allow
you to create a variable number of squares. createPad no longer accepts a number argument.
You will need to enter a number into the form and run the createPad function from the console
to make the grid at this point.